Enhance build & test system.
- `@nest/cli` for `npm run start:dev` command
- Refactor test program for re-usability
- Simultaneous client connections for rapid testing
- Adapt `typos` for spell checking
