+++ .github/dependabot.yml
@@ -6,7 +6,7 @@ updates:
     schedule:
       interval: "daily"
     open-pull-requests-limit: 25
-    # versioning-strategy: increase
+    versioning-strategy: increase
     allow:
       - dependency-name: "nestia"
       - dependency-name: "@nestjs/*"
