--- src/api/Primitive.ts
@@ -1,1 +1,0 @@
-export type { Primitive } from "@nestia/fetcher";
