--- src/api/Resolved.ts
@@ -1,1 +1,0 @@
-export type { Resolved } from "@nestia/fetcher";
