+++ src/api/module.ts
@@ -1,5 +1,4 @@
 export type * from "./IConnection";
-export type * from "./Primitive";
 export * from "./HttpError";
 
 export * as functional from "./functional";
